+++ RELEASE_NOTES.md
@@ -160,6 +160,7 @@ Note that a bunch of `StandardKeyboardLayoutProvider` things are deprecated in c
 * `InputSetProviderBased` is a new protocol that is used to keep track of types that rely on an input set provider.
 * `KeyboardActions` has new character margin actions properties.
 * `KeyboardContext` has a new `keyboardDictationReplacement` property.
+* `KeyboardEnabledLabel` is a new view that can be used to show enabled/active state.
 * `KKL10n` has a new `join` case, although localiations are missing for most locales.
 * `StandardInputSetProvider` `keyboardContext` is now public.
 * `StandardKeyboardLayoutProvider` can now take a collection of localized layout providers.